requirements.txt
scratch2py==0.6.0
requests==2.25.1
python-dotenv==0.19.1
